requirements.txt
absl-py==2.3.1
altair==5.4.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
easyocr==1.7.2
executing==2.2.1
fast-plate-ocr==1.0.1
fastapi==0.128.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.9.23
fonttools==4.61.1
fsspec==2025.12.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.3.1
humanfriendly==10.0
idna==3.11
imageio==2.37.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jax==0.8.2
jaxlib==0.8.2
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kagglehub==0.4.0
kagglesdk==0.1.14
keras==3.13.0
kiwisolver==1.4.9
lap==0.5.12
lazy_loader==0.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.18.0
packaging==24.2
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.3.0.post6
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.3
rich==13.9.4
rpds-py==0.30.0
scikit-image==0.24.0
scipy==1.16.3
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
starlette==0.50.0
streamlit==1.40.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tifffile==2025.12.12
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.140
ultralytics-thop==2.0.18
urllib3==2.6.2
uvicorn==0.40.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.5
wrapt==2.0.1
